chore: Add functional component wrapper and refactor
