+++ src/pages/introduction-apm.js
@@ -99,7 +99,7 @@ const IntroductionApm = () => {
   }, []);
 
   return (
-    <ErrorBoundary tessen={tessen} eventName="apm">
+    <ErrorBoundary eventName="apm">
       <FeaturedContent
         text={[t('apm.intro.text.0'), t('apm.intro.text.1')]}
         title={t('apm.intro.title')}
