+++ src/templates/embedPage.js
@@ -1,12 +1,7 @@
 import React, { useEffect } from 'react';
 import { graphql } from 'gatsby';
 import PropTypes from 'prop-types';
-import {
-  Layout,
-  useQueryParams,
-  Link,
-  useTessen,
-} from '@newrelic/gatsby-theme-newrelic';
+import { Layout, useQueryParams, Link } from '@newrelic/gatsby-theme-newrelic';
 import useDarkMode from 'use-dark-mode';
 import EmbedContext from '../components/EmbedContext';
 import ErrorBoundary from '../components/ErrorBoundary';
@@ -26,8 +21,6 @@ const EmbedPage = ({ data }) => {
   const embedDarkMode = queryParams.get('embedDarkMode');
   const darkMode = useDarkMode(embedDarkMode === 'enabled');
 
-  const tessen = useTessen();
-
   useEffect(() => {
     if (embedDarkMode === 'enabled') {
       darkMode.enable();
@@ -37,7 +30,7 @@ const EmbedPage = ({ data }) => {
   }, [darkMode, embedDarkMode]);
 
   return (
-    <ErrorBoundary tessen={tessen} eventName="embed">
+    <ErrorBoundary eventName="embed">
       <EmbedContext.Provider value={{ isEmbedded: true }}>
         <h1>{title}</h1>
         <Layout.Content>
